requirements.txt
numpy
tensorflow
streamlit
Pillow
opencv-python-headless